added getThumbnail
