+++ frontend/src/app/containers/CoverPage/index.tsx
@@ -6,6 +6,7 @@ import * as apiActions from 'api/actions';
 import { selectCover } from './slice/selectors';
 import WavePlayer from './WavePlayer';
 import * as urls from 'utils/urls';
+import { getThumbnail } from 'utils/functions';
 
 interface MatchParams {
   id?: string;
@@ -47,8 +48,7 @@ export default function CoverPage(props: Props) {
                 {cover.song.reference && (
                   <img
                     className="h-full w-full object-cover rounded-full text-center"
-                    // FIXME("need to use extract youtube thumnail link")
-                    src={cover.song.reference}
+                    src={getThumbnail(cover.song.reference)}
                     alt="Loading Thumbnail..."
                   />
                 )}
