+++ frontend/src/app/containers/MainPage/dummy.ts
@@ -1,45 +1,38 @@
-import image0 from '../../components/Album/AlbumImage0.jpg';
-import image1 from '../../components/Album/AlbumImage1.jpg';
-import image2 from '../../components/Album/AlbumImage2.jpg';
-import image3 from '../../components/Album/AlbumImage3.jpg';
-import image4 from '../../components/Album/AlbumImage4.jpg';
-import image5 from '../../components/Album/AlbumImage5.jpg';
-
 export const dummyAlbums = [
   {
     id: 0,
     title: '신호등',
     singer: '이무진',
-    thumbnail: image2,
+    reference: 'https://www.youtube.com/watch?v=SK6Sm2Ki9tI',
   },
   {
     id: 1,
     title: 'strawberry moon',
     singer: '아이유',
-    thumbnail: image1,
+    reference: 'https://www.youtube.com/watch?v=sqgxcCjD04s',
   },
   {
     id: 2,
     title: 'Savage',
     singer: 'aespa',
-    thumbnail: image0,
+    reference: 'https://www.youtube.com/watch?v=WPdWvnAAurg',
   },
   {
     id: 3,
     title: 'STAY',
     singer: 'The Kid LAROI, Justin Bieber',
-    thumbnail: image3,
+    reference: 'https://www.youtube.com/watch?v=rkYlZnIbe2E',
   },
   {
     id: 4,
     title: 'My Universe',
     singer: 'Coldplay, 방탄소년단',
-    thumbnail: image4,
+    reference: 'https://www.youtube.com/watch?v=bZYPI4mYwhw',
   },
   {
     id: 5,
     title: 'OHAYO MY NIGHT',
     singer: '디핵 (D-Hack), PATEKO (파테코)',
-    thumbnail: image5,
+    reference: 'https://www.youtube.com/watch?v=KCpWMEsiN3Q',
   },
 ];
